refactor: move governance out of the app crate to its component crate
