--- crates/core/app/src/governance/mod.rs
@@ -1,12 +1,0 @@
-mod component;
-pub(crate) mod event;
-mod metrics;
-mod view;
-
-pub mod state_key;
-pub mod tally;
-
-pub use self::metrics::register_metrics;
-pub use component::Governance;
-pub use tally::Tally;
-pub use view::{StateReadExt, StateWriteExt};
