+++ crates/core/app/src/lib.rs
@@ -8,7 +8,6 @@ pub use mock_client::MockClient;
 pub use temp_storage_ext::TempStorageExt;
 
 pub mod app;
-pub mod governance;
 
 #[cfg(test)]
 mod tests;
