+++ crates/core/component/governance/src/action_handler.rs
@@ -1,0 +1,1 @@
+use penumbra_component::ActionHandler;
