+++ crates/core/component/governance/src/delegator_vote.rs
@@ -1,1 +1,7 @@
+pub mod action;
+pub mod plan;
 pub mod proof;
+
+pub use action::{DelegatorVote, DelegatorVoteBody};
+pub use plan::DelegatorVotePlan;
+pub use proof::{DelegatorVoteCircuit, DelegatorVoteProof};
