+++ crates/core/component/governance/src/proposal_deposit_claim.rs
@@ -1,0 +1,3 @@
+pub mod action;
+
+pub use action::ProposalDepositClaim;
