+++ crates/core/component/governance/src/proposal_submit.rs
@@ -1,0 +1,3 @@
+pub mod action;
+
+pub use action::ProposalSubmit;
