+++ crates/core/component/governance/src/proposal_withdraw.rs
@@ -1,0 +1,3 @@
+pub mod action;
+
+pub use action::ProposalWithdraw;
