+++ crates/core/component/governance/src/validator_vote.rs
@@ -1,0 +1,3 @@
+pub mod action;
+
+pub use action::ValidatorVote;
