+++ crates/core/component/governance/src/vote.rs
