+++ crates/core/transaction/src/lib.rs
@@ -27,9 +27,7 @@ mod witness_data;
 pub mod action;
 pub mod memo;
 pub mod plan;
-pub mod proposal;
 pub mod view;
-pub mod vote;
 
 pub use action::Action;
 pub use auth_data::AuthorizationData;
